RP: Fix positioning of This Item
I changed the breakpoint to our medium size, since the images looked bad around 650-750px wide.

That fixed the issue with This Item almost filling the size of the image.

I adjusted the top positioning of This Item to match the padding at the top of its parent and then use margin to get its offset.