RP: remove shuffle + take 2 feature
We'll let the user of the component do this so the result
is stable instead of random everytime we change the props.

Specifically, if the caller flips this component between
several sets of products, it should be deterministic. I think
useMemo() only memoizes one value, not a map of them.