Deps: make react a peer and dev dependency
React and styled-components can't handle multiple copies of
the library being installed on one page. Therefore
it's recommended that these both be listed as peer dependencies
so the upstream user of this component controls which version
and ensures there's only one.

We keep them as devDependencies because we have an example
app that requires them.