Run `npm install` as `node` user
Installing as root seems to result in very strange errors.
`github:ifixit/feather` is a dependency of the npm-available
`@ifixit/toolbox` package which is one of our dependencies.  When `npm
install` is run as root, it seems to silently skip running the
`prepare` script from `ifixit/feather` which builds the icons. This
results in our `node_modules` not containing the build output, which
means our tests fail because they can't find their dependency files.

When `npm` is run as the `node` user (who exists by default in the
`node` image) we get all the dependencies installed correctly. The
`chown` here is needed because the `WORKDIR` won't get created with
the `node` user (regardless of whether the `USER` directive is given
before or after the `WORKDIR` directive).