Jest: compile less files than before
In c78a95599a795229f37b2053ef413dd8409d6cd2, we compiled some files in
node modules to fix a strange error. The solution I used was based on a
SO post here:
https://stackoverflow.com/questions/49263429/jest-gives-an-error-syntaxerror-unexpected-token-export

This was needed because Jest doesn't compile everything like webpack
does when we run `npm start` on this repo. A file in `ifixit/localize`
was using ES6 export syntax. When Jest tried to run this file, it failed
on this export statement (since it was never compiled). Let's change the
regex in `transformIgnorePatters` to only ignore the files in node
modules that aren't in the `ifixit/localize` repo.