npm install jest-styled-components
This is makes reading the diffs of snapshots easier to read if the diff
is regarding a changed style in a styled component.

Note: a previous version was installed to match our core components
package.json. Installing the newest version leads to a problem described
here: https://github.com/styled-components/jest-styled-components/issues/297