NotifyProduct: add new component
This component tries to match the one on iFixit.com within reason,
cleaning up a few minor spacing hacks that the original used. We use
Toolbox for the buttons and textbox, so these don't match the original
exactly.

We don't make an API call to actually prepare to notify the user, but
this will be addressed in a future commit.

Additionally, this component doesn't have any transitions like the
original, but this will added in a future commit.

On our example page, let's put this new component in a container that
simulates the product description container on iFixit.com and the
Shopify stores.