Confirmation: extract to status
If our API call fails, we're going to need an error dialogue. Let's
extract `Confirmation` into its own file and rename it status. We can add
a new prop, `type`, that we can use to conditionalize colors on.

I used the same error colors as the `UI/templates/status.phtml`
component in the main repo.