Notify confirmation: deal with error
We had no module handling API requests in this repo since no other
components were making API requests. This commit adds a simple method
for sending POST requests.

If we receive an error, we display the message of that error object
which might not be the best user message, but it'll do for now.