Notify: use email from input
Before, we were just sending the placeholder as the email in the request
body. With the `TextField` component we need to manage the state
ourselves, so this commit adds the `formEmail` state and sends that in
the request body.

We also disable the button until we have something to send. Ideally, we
would disable it until the text field is valid (a valid email since we
included `type="email"`) but we don't have a state for the validity of
the text field yet. Hopefully that can be added in a future commit.