Store dropdown: copy from main repo
This is a straight copy-paste with a few exceptions. Obviously the file
locations have changed so many of the imports needed to change. This also
added `box-sizing: border-box;` to `StyledCardLink` in
`src/lib/dropdowns/store/store_card.js` because our main repo adds
that style to all elements and this one does not. This was messing with
spacing of the example added in this commit. Also, the svg files for the
flags were converted to React components, but the contents are
untouched. We don't have webpack setup in this repo, so we couldn't use
`react-svg-loader`. Finally, this also passes through `className` to
`StoreDropdown` so we can use `styled-components` on it.

The example adds a container around the dropdown to try to mimic the
desktop dropdown on the live site.