Add jsonld data for product reviews
This allows google to see our review data better. We had to install
react helmet to be able to update the head from clientside.
Hopefully this resolves our search console issues.

We had to incorporate a workaround to get the product name as of now
as the data isn't easily provided