Inject jsonld prop into the html head
As of iFixit/ifixit#40347 we are creating this jsonld on the backend, so we don't need to build it in JS anymore.

If there is no jsonld available, the `jsonld` variable will be an empty object. In that case, we'll omit including the script tag entirely, because we have no jsonld to inject.